[feat][req]投料单领料批次锁定接口

+++ wip-spi/src/main/java/com/cvte/scm/wip/spi/requirement/dto/LotIssuedWriteBackDTO.java
@@ -22,7 +22,7 @@ public class LotIssuedWriteBackDTO {
 
     private String itemId;
 
-    private Integer lotQty;
+    private Long lotQty;
 
     private String opType;
